requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.16.0
psycopg2==2.9.8
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3

